LayoutTests/http/tests/inspector/network/x-frame-options-deny-expected.txt

CONSOLE MESSAGE: Refused to display document because display forbidden by X-Frame-Options.

Tests that responseReceived is called on NetworkDispatcher for resource requests denied due to X-Frame-Options header.


Received response for x-frame-options-deny.cgi
SUCCESS
